Restore Video#exist? method
Was failing due to api_uri being parsed as a generic URI which caused
problems when passed to Net::HTTP.get_response. Since the specs were
using stubs they didn't catch the problem.

We could probably add VCR to stub out the interactions but since this is
so simple I'm not sure it's even necessary.